add Makefill and debug

+++ ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: pnaessen <pnaessen@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/06 14:47:11 by pnaessen          #+#    #+#             */
-/*   Updated: 2024/11/11 08:12:20 by pnaessen         ###   ########lyon.fr   */
+/*   Updated: 2024/11/11 12:27:26 by pnaessen         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_bzero(void *s, size_t n)
 {
-	int				i;
+	size_t			i;
 	unsigned char	*temp;
 
 	i = 0;
